--- src/main/java/com/module/gomodules/controller/UserController.java
@@ -1,2 +1,0 @@
-package com.module.gomodules.controller;public class UserController {
-}
